apps/pyserver/requirements.txt
-i https://pypi.org/simple
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cycler==0.10.0
fastapi==0.63.0
h11==0.12.0; python_version >= '3.6'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
kiwisolver==1.3.1; python_version >= '3.6'
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
pillow==8.1.2
protobuf==3.15.6
pydantic==1.8.1; python_full_version >= '3.6.1'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-multipart==0.0.5
requests==2.25.1
six==1.15.0
starlette==0.13.6; python_version >= '3.6'
tensorflow-hub==0.11.0
typing-extensions==3.7.4.3
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.13.4
